% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_model_inputs.R
\name{extract_ip}
\alias{extract_ip}
\title{extract_IP
Function to extract infestation pressure and sea temperature for a farm
based on the location ID.}
\usage{
extract_ip(lokid)
}
\arguments{
\item{lokid}{The Location ID}
}
\description{
The extraction is done from internal data. These needs to be updated when
changes occur. See code in data-raw for more details.
}
\examples{
\dontrun{
}
}
