% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_model_inputs.R
\name{extract_lice}
\alias{extract_lice}
\title{extract_lice}
\usage{
extract_lice(lokid)
}
\arguments{
\item{lokid}{is the location ID}
}
\value{
The output is a list with the following elements:
AF is the latest reported number of adult female lice per fish in the farm;
OM is the latest reported number of other motile lice per fish in the farm;
FX is the latest reported number of sessile lice per fish in the farm.
}
\description{
Function to extract lice data for a farm based on its location ID.
}
\details{
The extraction is done from internal data. These needs to be updated when
changes occur. See code in data-raw for more details.
}
\examples{
#' \dontrun{
}
}
