% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_model.R
\name{plot_lice}
\alias{plot_lice}
\title{plot_lice
Function to plot predictions for one cage}
\usage{
plot_lice(prediction.object, cage = 1, lan = "no")
}
\arguments{
\item{prediction.object}{is a prediction object created with the predict_lice function}

\item{cage}{is the cage number}

\item{lan}{is language ('no' = Norwegian, 'en' = English)}
}
\value{
The output is a prediction plot for one cage
}
\description{
plot_lice
Function to plot predictions for one cage
}
\examples{
\dontrun{
AF = 1:3
OM = 3:1
FX = c(0,0,.5)
prediction_object = predict_lice(AF=AF, OM=OM, FX=FX)
par(mai = c(.7,.7,.6,.01))
plot_lice(prediction.object, cage = 2, lan="en")
}
}
