% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_model.R
\name{plot_prediction}
\alias{plot_prediction}
\title{plot_predition
A function to generate plot from the predicted data.}
\usage{
plot_prediction(prediction_df, log = TRUE)
}
\arguments{
\item{prediction_df}{a dataframe returned by `restult_to_df`}

\item{log}{flag for converting the data to log scale. `TRUE` by default.}
}
\value{
a ggplot2 object
}
\description{
plot_predition
A function to generate plot from the predicted data.
}
\examples{
\dontrun{
}
}
