% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_lice.R
\name{plotLok}
\alias{plotLok}
\title{plotLok}
\usage{
plotLok(loknr, valgtSm, tid, innlestData, plotnr, hele, merd)
}
\arguments{
\item{loknr}{description}
}
\value{
A plot object
}
\description{
A function to plot lice development prediction
}
