% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_predict_lice.R
\name{predict_lice}
\alias{predict_lice}
\title{predict_lice
Function to predict lice per fish in different stages and cages one to two weeks ahead}
\usage{
predict_lice(
  AF,
  OM,
  FX,
  ST = 9,
  W_SAL = 2.2,
  N_SAL = 130,
  CLF = F,
  IP_1wk = 94000,
  IP_2wk = 94000,
  q1 = 0.05,
  q2 = 0.95,
  ncount = 20
)
}
\arguments{
\item{AF}{vector with number of adult female lice per fish in each cage in the farm}

\item{OM}{vector with number of other motile lice per fish in each cage in the farm}

\item{FX}{vector with number of fixed lice per fish in each cage in the farm}

\item{ST}{sea temperature (degrees C)}

\item{W_SAL}{mean salmon weight (kg) in each cage}

\item{N_SAL}{salmon abundance (thousands) in each cage}

\item{CLF}{presence or absence of sea wrasses as cleaner fish in each cage}

\item{IP_1wk}{external lice infestation pressure for FX one week ahead}

\item{IP_2wk}{external lice infestation pressure for FX two weeks ahead}

\item{q1}{lower limit of prediction interval}

\item{q2}{upper limit of prediction interval}

\item{ncount}{number of salmon counted in each cage to be predicted}
}
\value{
The output is a list with the following elements:
\itemize{
  \item FX_1wk is a list with predictions for the number of sessile lice 1 week ahead
  \item OM_1wk is a list with predictions for the number of other motile lice 1 week ahead
  \item AF_1wk is a list with predictions for the number of adult female lice 1 week ahead
  \item FX_2wk is a list with predictions for the number of sessile lice 2 weeks ahead
  \item OM_2wk is a list with predictions for the number of other motile lice 2 weeks ahead
  \item AF_2wk is a list with predictions for the number of adult female lice 2 weeks ahead
  }
Each of these list elements contains the following elements:
\itemize {
  \item mu is the expected number of lice per fish
  \item ci1_true is the lower limit for the true number of lice per fish
  \item ci2_true is the upper limit for the true number of lice per fish
  \item ci1_count is the lower limit for the counted number of lice per fish
  \item ci2_count is the upper limit for the counted number of lice per fish
  }
}
\description{
predict_lice
Function to predict lice per fish in different stages and cages one to two weeks ahead
}
\examples{
\dontrun{
AF = 1:3
OM = 3:1
FX = c(0,0,.5)
prediction_object = predict_lice(AF=AF, OM=OM, FX=FX)
}
}
