% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_model.R
\name{restult_to_df}
\alias{restult_to_df}
\title{results_to_df}
\usage{
restult_to_df(prediction_object, log = TRUE)
}
\arguments{
\item{prediction_object}{a list of the predicted values returned by the
prediction fucntion}

\item{log}{flag for converting the data to log scale. `TRUE` by default.}
}
\value{
a dataframe of the prediction
}
\description{
This function is used to convert list returned from the prediction
function into a data frame for easier manipulation and plotting.
}
\details{
The function additionally transforms the data to log with `ytrans`.
}
\examples{
\dontrun{
}
}
